% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{claim_id}
\alias{claim_id}
\title{Identify Claims in Text}
\usage{
claim_id(df, text_col, search_word, synon = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{text_col}{The name of the text column}

\item{search_word}{The word to search for}

\item{synon}{Boolean indicating whether to use synonyms}
}
\value{
Data frame with an "identifier" column
}
\description{
Identify Claims in Text
}
