% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classify_representative_claims}
\alias{classify_representative_claims}
\title{Classify Representative Claims}
\usage{
classify_representative_claims(
  df,
  text_col,
  label_col,
  training_df,
  training_text_col,
  method
)
}
\arguments{
\item{df}{A data frame containing the text to classify}

\item{text_col}{The column name containing text}

\item{label_col}{The column name containing labels for training}

\item{training_df}{A data frame containing training data}

\item{training_text_col}{The column name containing text in the training data}

\item{method}{The classification method to use ("svm", "naive_bayes", "xlnet", "xlm-roberta", "mbert")}
}
\value{
A data frame with classification results
}
\description{
Classify Representative Claims
}
