% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classify_text}
\alias{classify_text}
\title{Classify text using a trained model}
\usage{
classify_text(df, text_col, model)
}
\arguments{
\item{df}{Data frame with text to classify}

\item{text_col}{Column name containing text}

\item{model}{Trained model (SVM or Naïve Bayes)}
}
\value{
Data frame with predictions
}
\description{
Classify text using a trained model
}
