% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi_unrestricted_variable_appraisals.R
\name{gi_unrestricted_variable_appraisals.rs_singlestage}
\alias{gi_unrestricted_variable_appraisals.rs_singlestage}
\title{Get Input Information for RAT-STATS Style Unrestricted Variable Appraisals}
\usage{
\method{gi_unrestricted_variable_appraisals}{rs_singlestage}(
  samp_obj,
  data_file_format = c("Audited Values"),
  audited_values = NULL,
  sample_item_number
)
}
\arguments{
\item{samp_obj}{A nested list returned from a claimr sampling function}

\item{data_file_format}{"Data File Format" for RAT-STATS extrapolation}

\item{audited_values}{Column with "number being reviewed by the user"}

\item{sample_item_number}{Column with "sample item number"}
}
\value{
Information and dataset required for a RAT-STATS Unrestricted Variable Appraisals
}
\description{
This function takes a nested list from a claimr sampling function and returns
the information needed to perform a RAT-STATS' Unrestricted Variable Appraisals
}
\examples{
#Generate sampling frame  with an index id from 1 to N
math_test <- data.frame("sample_frame_sequence_id" = 1:1000,
                        "math" = rnorm(1000, mean = 75, sd = 15))

#Get output and input in lists
math_test_cobj <- rs_singlestage(df = math_test,
                                 seed_number = 100,
                                 audit_review = "Math Sample",
                                 quantity_to_generate = 20,
                                 quantity_of_spares = 5,
                                 frame_low = 1,
                                 frame_high = 1000)

gi_unrestricted_variable_appraisals(samp_obj = math_test_cobj,
                                    data_file_format = c("Audited Values"),
                                    audited_values = math,
                                    sample_item_number = sample_frame_sequence_id)
}
