% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_sum_method.R
\name{Ne_low_minus_ne}
\alias{Ne_low_minus_ne}
\title{Difference Between Ne_low ("lower bound for the number of population payments in error") and ne}
\usage{
Ne_low_minus_ne(Npop, nsamp, ne, alpha = 0.1)
}
\arguments{
\item{Npop}{"Number of payments (e.g. on Medicare claims) in the universe/population"}

\item{nsamp}{"Number of payments (e.g. on Medicare claims) in...simple random sample"}

\item{ne}{"The number...of sample payments which are completely in error (Or, for partial overpayment scenarios, seriously in error—see Section 4.2.)"}

\item{alpha}{alpha-level for "1-alpha confidence bound"}
}
\value{
Ne_low - ne
}
\description{
Difference Between Ne_low ("lower bound for the number of population payments in error") and ne
}
\examples{
Ne_low_minus_ne(Npop = 1000, nsamp = 30, ne = 20)
}
