% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_sum_method.R
\name{Ne_up}
\alias{Ne_up}
\title{Upper 1-alpha Confidence Bound (Minimum Sum Method)}
\usage{
Ne_up(Npop, nsamp, ne, alpha = 0.1)
}
\arguments{
\item{Npop}{"Number of payments (e.g. on Medicare claims) in the universe/population"}

\item{nsamp}{"Number of payments (e.g. on Medicare claims) in...simple random sample"}

\item{ne}{"The number...of sample payments which are completely in error (Or, for partial overpayment scenarios, seriously in error—see Section 4.2.)"}

\item{alpha}{alpha-level for "1-alpha confidence bound"}
}
\value{
Upper 1-alpha confidence bound for the "number...of universe payments which are completely in error (Or, for partial overpayment scenarios, seriously in error—see Section 4.2.)"
}
\description{
Function obtained from Edwards et. al. (2003): "a bisection routine to find an upper 1-alpha confidence bound for Ne, the number of population items in error. Population size Npop, SRS size nsamp, ne = number of sample items in error."
}
\details{
Citation: Edwards, D., Ward-Besser, G., Lasecki, J., Parker, B., Wu, F. & Moorhead, P. (2003).  The Minimum Sum Method: A Distribution-Free Sampling Procedure for Medicare Fraud Investigations.  Heath Services & Outcomes Research Methodology 4: 241-263.
}
\examples{
Ne_up(Npop = 1000, nsamp = 30, ne = 15, alpha=0.05)
}
