% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_setsoftwo.R
\name{rs_setsoftwo}
\alias{rs_setsoftwo}
\title{Rat-Stats Style Sets of Two Random Numbers}
\usage{
rs_setsoftwo(
  df = NULL,
  first_set,
  second_set,
  seed_number = NA,
  audit_review = "",
  quantity_to_generate = 0,
  quantity_of_spares = 0,
  first_set_low = 1,
  first_set_high = 1,
  second_set_low = 1,
  second_set_high = 1
)
}
\arguments{
\item{df}{A data frame containing an index for the first and second set of numbers}

\item{first_set}{Index for first set of numbers (i.e., page, source, or data frame id)}

\item{second_set}{Index for second set of numbers (i.e., line, observation, or row id)}

\item{seed_number}{"Do you want to provide seed number?" Provide a seed number.
If the default (seed_number = NA) is retained then the seed number will be set
to sample(0:1000000, 1) and the result will be returned in the output.}

\item{audit_review}{"Name of the Audit/Review"}

\item{quantity_to_generate}{"Enter the quantity of numbers to be generated in
Random Order."}

\item{quantity_of_spares}{"Enter the quantity of Spare numbers to be generated
in Random Order:"}

\item{first_set_low}{Enter the low number for the sampling frame's first set}

\item{first_set_high}{Enter the high number for the sampling frame's first set}

\item{second_set_low}{Enter the low number for the sampling frame's second set}

\item{second_set_high}{Enter the high number for the sampling frame's second set}
}
\value{
A nested list with "output" and "input" lists.  The output list contains three
data frames: the sample, the sample frame, and the spares.  Each data frame
includes the random numbers (in the order variable) and each has been sorted by first_set and second_set.  The input list
returns input parameters and the random number seed.
}
\description{
Rat-Stats Style Sets of Two Random Numbers
}
\examples{
#Create three separate data frames
df1 <- data.frame(page = c(1),
                 score = rnorm(10, 7,2),
                 item = 1:10)
df2 <- data.frame(page = c(2),
                 score = rnorm(10, 6, 1.5),
                 item = 1:10)
df3 <- data.frame(page = c(3),
                 score = rnorm(10, 8, 0.5),
                 item = 1:10)

#Combine the data frames
df_combined <- rbind(df1, df2, df3)
#Randomly pull observations from df_combined
combined_out <- rs_setsoftwo(df = df_combined,
                            first_set = page,
                            second_set = item,
                            seed_number = NA,
                            audit_review = "",
                            quantity_to_generate = 10,
                            quantity_of_spares = 2,
                            first_set_low = 1,
                            first_set_high = 3,
                            second_set_low = 1,
                            second_set_high = 10)

#View head
head(combined_out$output$sample)
}
