% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_claimr}
\alias{run_claimr}
\title{ClaimR - Finding Representative Claims in TXT}
\usage{
run_claimr(df, text_col)
}
\arguments{
\item{df}{The data frame containing the claims.}

\item{text_col}{The name of the column containing the text (as a string).}
}
\description{
ClaimR - Finding Representative Claims in TXT
}
