% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_sentences}
\alias{split_sentences}
\title{Split Text into Sentences}
\usage{
split_sentences(df, text_col, circ = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{text_col}{The name of the text column in the data frame}

\item{circ}{A boolean indicating whether to include surrounding sentences}
}
\value{
A data frame with sentence IDs
}
\description{
Split Text into Sentences
}
